Check existing valid certificate to avoid duplicate cert requests.

+++ init.sh
@@ -30,43 +30,54 @@ if [ ! -e "$data_path/conf/options-ssl-nginx.conf" ] || [ ! -e "$data_path/conf/
   echo
 fi
 
-echo "### Creating dummy certificate for $domains ..."
-path="/etc/letsencrypt/live/$domains"
-mkdir -p "$data_path/conf/live/$domains"
-docker-compose run --rm --entrypoint "\
-  openssl req -x509 -nodes -newkey rsa:1024 -days 1\
-    -keyout '$path/privkey.pem' \
-    -out '$path/fullchain.pem' \
-    -subj '/CN=localhost'" certbot
-echo
+echo "### Checking if there is a valid certificate to avoid request limit..."
+days=$(( ($(date --date="$(openssl x509 -in ./data/certbot/conf/live/$domains/cert.pem -noout -dates  | grep notAfter | awk '{print $1,$2,$4}'| cut -b 10-14,15-)" "+%s") - $(date '+%s')) / 86400))
+cn=$(openssl x509 -noout -in ./data/certbot/conf/live/$domains/cert.pem -subject | cut -b 14-)
+if [ $cn = $domains ] && [ $days > 10 ]; then
+  echo "No need for new certificate"
+  no_new_cert=true
+fi
 
+
+if [ $no_new_cert != true ]; then
+  echo "### Creating dummy certificate for $domains ..."
+  path="/etc/letsencrypt/live/$domains"
+  mkdir -p "$data_path/conf/live/$domains"
+  docker-compose run --rm --entrypoint "\
+    openssl req -x509 -nodes -newkey rsa:1024 -days 1\
+      -keyout '$path/privkey.pem' \
+      -out '$path/fullchain.pem' \
+      -subj '/CN=localhost'" certbot
+  echo
+fi
 
 echo "### Starting nginx ..."
 docker-compose up --force-recreate -d nginx
 echo
 
-echo "### Deleting dummy certificate for $domains ..."
-docker-compose run --rm --entrypoint "\
-  rm -Rf /etc/letsencrypt/live/$domains && \
-  rm -Rf /etc/letsencrypt/archive/$domains && \
-  rm -Rf /etc/letsencrypt/renewal/$domains.conf" certbot
-echo
+if [ $no_new_cert != true ]; then
+  echo "### Deleting dummy certificate for $domains ..."
+  docker-compose run --rm --entrypoint "\
+    rm -Rf /etc/letsencrypt/live/$domains && \
+    rm -Rf /etc/letsencrypt/archive/$domains && \
+    rm -Rf /etc/letsencrypt/renewal/$domains.conf" certbot
+  echo
 
 
-echo "### Requesting Let's Encrypt certificate for $domains ..."
-domain_args="-d $domains"
+  echo "### Requesting Let's Encrypt certificate for $domains ..."
+  domain_args="-d $domains"
 
-# Select appropriate email arg
-case "$email" in
-  "") email_arg="--register-unsafely-without-email" ;;
-  *) email_arg="--email $email" ;;
-esac
+  # Select appropriate email arg
+  case "$email" in
+    "") email_arg="--register-unsafely-without-email" ;;
+    *) email_arg="--email $email" ;;
+  esac
 
-# Enable staging mode if needed
-if [ $staging != "0" ]; then staging_arg="--staging"; fi
+  # Enable staging mode if needed
+  if [ $staging != "0" ]; then staging_arg="--staging"; fi
 
-docker-compose run --rm --entrypoint "\
-  certbot certonly --webroot -w /var/www/certbot \
+  docker-compose run --rm --entrypoint "\
+    certbot certonly --webroot -w /var/www/certbot \
     $staging_arg \
     $email_arg \
     $domain_args \
@@ -75,11 +86,13 @@ docker-compose run --rm --entrypoint "\
     --noninteractive \
     --no-eff-email \
     --force-renewal" certbot
-echo
+  echo
 
-echo "### Reloading nginx ..."
-docker-compose exec nginx nginx -s reload
-echo 
+  echo "### Reloading nginx ..."
+  docker-compose exec nginx nginx -s reload
+  echo 
+fi
+
 
 echo "### Starting certbot in renew mode ..."
 docker-compose up -d certbot
